arm7: merge SND_wave into SND_exChannel
