+++ arm7/arm7.lsf
@@ -46,7 +46,6 @@ Autoload WRAM
 	Object SND_capture.o
 	Object SND_channel.o
 	Object SND_lfo.o
-	Object SND_wave.o
 	Object SND_exChannel.o
 	Object SND_seq.o
 	Object SND_bank.o
