--- arm7/lib/include/SND_wave.h
@@ -1,6 +1,0 @@
-#ifndef GUARD_SND_WAVE_H
-#define GUARD_SND_WAVE_H
-
-void SND_InvalidateWave(const void *begin, const void *end);
-
-#endif //GUARD_SND_WAVE_H
